[change] move page fetcher helpers to shared
